==> 1-PBMC-sc/01-qc.R <==

## 设置环境
rm(list = ls()) 
options(stringsAsFactors = F) 
library(Seurat)
library(ggplot2)
library(clustree)
library(cowplot)
library(dplyr)
library(stringr)
library(data.table)
library(ggsci)
load(file = "./0-Dataprepare/sce.all.Rdata")
load(file = "./0-Dataprepare/mycol.Rdata")
# sce.all = subset(sce.all, GSE == "GSE226602")

## 计算线粒体基因比例
mito_genes = rownames(sce.all)[grep("^MT-", rownames(sce.all))] # 人和鼠的基因名字稍微不一样 
mito_genes #13个线粒体基因
sce.all = PercentageFeatureSet(sce.all, "^MT-", col.name = "percent_mito")

## 计算核糖体基因比例
ribo_genes = rownames(sce.all)[grep("^RP[SL]", rownames(sce.all))]
ribo_genes
sce.all = PercentageFeatureSet(sce.all, "^RP[SL]", col.name = "percent_ribo")

## 计算红血细胞基因比例
rownames(sce.all)[grep("^HB[^(P)]", rownames(sce.all))]
sce.all = PercentageFeatureSet(sce.all, "^HB[^(P)]", col.name = "percent_hb")

## 可视化细胞的上述比例情况
feats = c("nFeature_RNA", "nCount_RNA")
VlnPlot(sce.all, group.by = "orig.ident", features = feats, pt.size = 0, ncol = 2, cols = col_vec) + NoLegend()

feats = c("percent_mito", "percent_ribo", "percent_hb")
VlnPlot(sce.all, group.by = "orig.ident", features = feats, pt.size = 0, ncol = 3, same.y.lims = T, cols = col_vec) + NoLegend()


# 根据上述指标，过滤低质量细胞/基因
## 过滤指标1:最少表达基因数的细胞&最少表达细胞数的基因
sce.all = JoinLayers(sce.all)
selected_c = WhichCells(sce.all, expression = (nFeature_RNA > 500) & (nFeature_RNA < 3000) & (nCount_RNA < 10000))
selected_f = rownames(sce.all)[Matrix::rowSums(sce.all@assays$RNA@layers$counts > 0) > 200]
sce.all.filt = subset(sce.all, features = selected_f, cells = selected_c)

dim(sce.all) 
dim(sce.all.filt) 

table(sce.all@meta.data$orig.ident) 
table(sce.all.filt@meta.data$orig.ident) 

## 过滤指标2:线粒体/核糖体基因比例(根据上面的violin图)
selected_mito = WhichCells(sce.all.filt, expression = percent_mito < 10)
selected_ribo = WhichCells(sce.all.filt, expression = percent_ribo > 3)
selected_hb = WhichCells(sce.all.filt, expression = percent_hb < 1)

length(selected_mito)
length(selected_ribo)
length(selected_hb)

sce.all.filt = subset(sce.all.filt, cells = selected_mito)
sce.all.filt = subset(sce.all.filt, cells = selected_ribo)
sce.all.filt = subset(sce.all.filt, cells = selected_hb)
dim(sce.all.filt)

table(sce.all.filt$orig.ident) 

## 过滤特定基因
sce.all.filt = sce.all.filt[!grepl("^MT-", rownames(sce.all.filt), ignore.case = T), ]
sce.all.filt = sce.all.filt[!grepl("^RP[SL]", rownames(sce.all.filt), ignore.case = T), ]
dim(sce.all.filt) 


## 可视化过滤后的情况
feats = c("nFeature_RNA", "nCount_RNA")
VlnPlot(sce.all.filt, group.by = "orig.ident", features = feats, pt.size = 0, ncol = 1, cols = col_vec) + NoLegend()
ggsave(filename="./1-qc/1-Filter_Vlnplot1.pdf", width = 13, height = 8)

feats = c("percent_mito", "percent_hb")
VlnPlot(sce.all.filt, group.by = "orig.ident", features = feats, pt.size = 0, ncol = 1, cols = col_vec) + NoLegend()
ggsave(filename="./1-qc/2-Filter_Vlnplot2.pdf", width = 13, height = 8)

## 保存数据
save(sce.all.filt, file = "./1-qc/sce.all.filt.Rdata")
